pyproject.toml
[tool.poetry]
name = "invest2iis"
version = "0.1.0"
description = "A Telegram bot for monitoring investment account status"
authors = ["falant@ya.ru"]

[tool.poetry.dependencies]
telebot = "0.0.5"
python-dotenv = "^1.0.1"
tinkoff-investments = "^0.0.0"

[tool.poetry.scripts]
start = "main:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"